Gets assignment of agent to threads table
