+++ backend/src/constants/__init__.py
@@ -25,6 +25,7 @@ JWT_ALGORITHM = "HS256"
 JWT_TOKEN_EXPIRE_MINUTES = 60 * 24
 
 # App
+APP_ENV = os.getenv("APP_ENV", "development")
 APP_VERSION = os.getenv("APP_VERSION", "0.1.0")
 APP_SECRET_KEY = os.getenv("APP_SECRET_KEY", "this-is-a-secret-key")
 APP_LOG_LEVEL = os.getenv("APP_LOG_LEVEL", "INFO").upper()
